+++ sqlmesh_dag_generator/__init__.py
@@ -2,7 +2,7 @@
 SQLMesh DAG Generator - Airflow orchestration for SQLMesh projects, self-hosted.
 """
 
-__version__ = "0.10.0"
+__version__ = "0.10.1"
 
 
 from sqlmesh_dag_generator.airflow_compat import (